fix signals test
